requirements.txt
python-dotenv
pipecat-ai[cartesia,openai,silero,websocket,deepgram,mem0]
uvicorn
fastapi
plivo
pydantic==2.10.4
